back/requirements.txt
bidict==0.21.2
click==8.0.1
Flask==2.0.1
Flask-SocketIO==5.1.0
h11==0.12.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
python-datauri==1.0.0
python-engineio==4.2.0
python-socketio==5.3.0
simple-websocket==0.2.0
Werkzeug==2.0.1
wsproto==1.0.0
